+++ datatransfer/impl/graphsync/graphsync_impl.go
@@ -1,15 +1,20 @@
 package graphsyncimpl
 
 import (
-	"bytes"
+	"context"
+	"fmt"
+	"reflect"
 
-	logging "github.com/ipfs/go-log"
+	"github.com/ipfs/go-cid"
+	"github.com/ipfs/go-graphsync"
 	"github.com/ipld/go-ipld-prime"
-	"github.com/ipld/go-ipld-prime/encoding/dagcbor"
-	ipldfree "github.com/ipld/go-ipld-prime/impl/free"
-)
+	"github.com/libp2p/go-libp2p-core/host"
+	"github.com/libp2p/go-libp2p-core/peer"
 
-var log = logging.Logger("graphsync-impl")
+	"github.com/filecoin-project/lotus/datatransfer"
+	"github.com/filecoin-project/lotus/datatransfer/message"
+	"github.com/filecoin-project/lotus/datatransfer/network"
+)
 
 const (
 	// ExtensionDataTransfer is the identifier for the data transfer extension to graphsync
@@ -42,11 +47,6 @@ type graphsyncImpl struct {
 	gs                  graphsync.GraphExchange
 	peerID              peer.ID
 	lastTID             int64
-}
-
-type graphsyncReceiver struct {
-	ctx  context.Context
-	impl *graphsyncImpl
 }
 
 // NewGraphSyncDataTransfer initializes a new graphsync based data transfer manager
@@ -100,7 +100,7 @@ func (impl *graphsyncImpl) OpenPushDataChannel(ctx context.Context, requestTo pe
 	if err != nil {
 		return datatransfer.ChannelID{}, err
 	}
-	// initiator = me, sender = me, receiver = them
+	// initiator = us, sender = us, receiver = them
 	chid := impl.createNewChannel(tid, baseCid, selector, voucher,
 		//requestTo, impl.peerID, requestTo)
 		impl.peerID, impl.peerID, requestTo)
@@ -115,15 +115,15 @@ func (impl *graphsyncImpl) OpenPullDataChannel(ctx context.Context, requestTo pe
 	if err != nil {
 		return datatransfer.ChannelID{}, err
 	}
-	// initiator = me, sender = them, receiver = me
+	// initiator = us, sender = them, receiver = us
 	chid := impl.createNewChannel(tid, baseCid, selector, voucher,
 		impl.peerID, requestTo, impl.peerID)
 	return chid, nil
 }
 
 // createNewChannel creates a new channel id and channel state and saves to channels
-func (impl *graphsyncImpl) createNewChannel(tid datatransfer.TransferID, baseCid cid.Cid, selector ipld.Node, voucher datatransfer.Voucher, initiator, dataSender, dataReceiver peer.ID) datatransfer.ChannelID {
-	chid := datatransfer.ChannelID{Initiator: initiator, ID: tid}
+func (impl *graphsyncImpl) createNewChannel(tid datatransfer.TransferID, baseCid cid.Cid, selector ipld.Node, voucher datatransfer.Voucher, requestTo, dataSender, dataReceiver peer.ID) datatransfer.ChannelID {
+	chid := datatransfer.ChannelID{Initiator: requestTo, ID: tid}
 	chst := datatransfer.ChannelState{Channel: datatransfer.NewChannel(0, baseCid, selector, voucher, dataSender, dataReceiver, 0)}
 	impl.channels[chid] = chst
 	return chid
@@ -196,106 +196,6 @@ func (impl *graphsyncImpl) InProgressChannels() map[datatransfer.ChannelID]datat
 	return impl.channels
 }
 
-// ReceiveRequest takes an incoming request, validates the voucher and processes the message.
-func (receiver *graphsyncReceiver) ReceiveRequest(
-	ctx context.Context,
-	sender peer.ID,
-	incoming message.DataTransferRequest) {
-
-	// not yet doing anything else with the voucher
-	_, err := receiver.validateVoucher(sender, incoming)
-	if err != nil {
-		receiver.impl.sendResponse(ctx, false, sender, incoming.TransferID())
-		return
-	}
-	stor, _ := nodeFromBytes(incoming.Selector())
-	root := cidlink.Link{incoming.BaseCid()}
-	if !incoming.IsPull() {
-		receiver.impl.gs.Request(ctx, sender, root, stor)
-	}
-	receiver.impl.sendResponse(ctx, true, sender, incoming.TransferID())
-}
-
-// validateVoucher converts a voucher in an incoming message to its appropriate
-// voucher struct, then runs the validator and returns the results.
-// returns error if:
-//   * voucherFromRequest fails
-//   * deserialization of selector fails
-//   * validation fails
-func (receiver *graphsyncReceiver) validateVoucher(sender peer.ID, incoming message.DataTransferRequest) (datatransfer.Voucher, error) {
-
-	vtypStr := incoming.VoucherType()
-	vouch, err := receiver.voucherFromRequest(incoming)
-	if err != nil {
-		return vouch, err
-	}
-
-	var validatorFunc func(peer.ID, datatransfer.Voucher, cid.Cid, ipld.Node) error
-	if incoming.IsPull() {
-		validatorFunc = receiver.impl.validatedTypes[vtypStr].validator.ValidatePull
-	} else {
-		validatorFunc = receiver.impl.validatedTypes[vtypStr].validator.ValidatePush
-	}
-
-	stor, err := nodeFromBytes(incoming.Selector())
-	if err != nil {
-		return vouch, err
-	}
-
-	if err = validatorFunc(sender, vouch, incoming.BaseCid(), stor); err != nil {
-		return nil, err
-	}
-
-	return vouch, nil
-}
-
-// voucherFromRequest takes an incoming request and attempts to create a
-// voucher struct from it using the registered validated types.  It returns
-// a deserialized voucher and any error.  It returns error if:
-//    * the voucher type has no validator registered
-//    * the voucher cannot be instantiated via reflection
-//    * request voucher bytes cannot be deserialized via <voucher>.FromBytes()
-func (receiver *graphsyncReceiver) voucherFromRequest(incoming message.DataTransferRequest) (datatransfer.Voucher, error) {
-	vtypStr := incoming.VoucherType()
-
-	validatedType, ok := receiver.impl.validatedTypes[vtypStr]
-	if !ok {
-		return nil, fmt.Errorf("unregistered voucher type %s", vtypStr)
-	}
-	vStructVal := reflect.New(validatedType.voucherType.Elem())
-	voucher, ok := vStructVal.Interface().(datatransfer.Voucher)
-	if !ok || reflect.ValueOf(voucher).IsNil() {
-		return nil, fmt.Errorf("problem instantiating type %s, voucher: %v", vtypStr, voucher)
-	}
-	if err := voucher.FromBytes(incoming.Voucher()); err != nil {
-		return voucher, err
-	}
-	return voucher, nil
-}
-
-// ReceiveResponse handles responding to Push or Pull Requests.
-// It schedules a graphsync transfer only if a Pull Request is accepted.
-func (receiver *graphsyncReceiver) ReceiveResponse(
-	ctx context.Context,
-	sender peer.ID,
-	incoming message.DataTransferResponse) {
-	evt := datatransfer.Error
-	chst := datatransfer.EmptyChannelState
-	if incoming.Accepted() {
-		chid := datatransfer.ChannelID{
-			Initiator: receiver.impl.peerID,
-			ID:        incoming.TransferID(),
-		}
-		if chst = receiver.impl.getPullChannel(chid); chst != datatransfer.EmptyChannelState {
-			baseCid := chst.BaseCID()
-			root := cidlink.Link{baseCid}
-			receiver.impl.gs.Request(ctx, sender, root, chst.Selector())
-			evt = datatransfer.Progress
-		}
-	}
-	receiver.impl.notifySubscribers(evt, chst)
-}
-
 // getPullChannel searches for a pull-type channel in the slice of channels with id `chid`.
 // Returns datatransfer.EmptyChannelState if:
 //   * there is no channel with that id
@@ -308,26 +208,8 @@ func (impl *graphsyncImpl) getPullChannel(chid datatransfer.ChannelID) datatrans
 	return channelState
 }
 
-func (receiver *graphsyncReceiver) ReceiveError(error) {}
-
-// nodeAsBytes serializes an ipld.Node
-func nodeAsBytes(node ipld.Node) ([]byte, error) {
-	var buffer bytes.Buffer
-	err := dagcbor.Encoder(node, &buffer)
-	if err != nil {
-		return nil, err
-	}
-	return buffer.Bytes(), nil
-}
-
-// nodeFromBytes deserializes an ipld.Node
-func nodeFromBytes(from []byte) (ipld.Node, error) {
-	reader := bytes.NewReader(from)
-	return dagcbor.Decoder(ipldfree.NodeBuilder(), reader)
-}
-
-// generateTransferID() generates a unique-to-runtime TransferID for use in creating
-// ChannelIDs
+// TODO: implement a real transfer ID generator.
+// https://github.com/filecoin-project/go-data-transfer/issues/38
 func (impl *graphsyncImpl) generateTransferID() datatransfer.TransferID {
 	impl.lastTID++
 	return datatransfer.TransferID(impl.lastTID)
